+++ pyproject.toml
@@ -91,7 +91,7 @@ ps_search = 'python_search.search.search_entries:main'
 ps_semantic_search = 'python_search.semantic_search.text2embeddings:main'
 ps_text_next_predictor = 'python_search.textual_next_predictor.predictor:main'
 start_streamlit_app = 'start_streamlit_app:main'
-llm_rank = 'python_search.llm_next_item_predictor.t5:main'
+llm_rank = 'python_search.llm_next_item_predictor.t5.inference:main'
 
 [tool.poetry.extras]
 model = [
